+++ src/Products/MailHost/tests/test_exportimport.py
@@ -15,6 +15,8 @@
 
 import unittest
 
+import zope.component
+
 try:
     from Products.GenericSetup.testing import BodyAdapterTestCase
     from Products.GenericSetup.testing import ExportImportZCMLLayer
@@ -62,6 +64,11 @@ class MailHostXMLAdapterTests(BodyAdapterTestCase, unittest.TestCase):
 
         self._obj = MailHost('foo_mailhost')
         self._BODY = _MAILHOST_BODY
+        zope.component.provideAdapter(self._getTargetClass())
+
+    def tearDown(self):
+        gsm = zope.component.getGlobalSiteManager()
+        self.assertTrue(gsm.unregisterAdapter(self._getTargetClass()))
 
 
 @unittest.skipUnless(HAVE_GS, 'Products.GenericSetup not available.')
@@ -78,6 +85,7 @@ class MailHostXMLAdapterTestsWithoutQueue(MailHostXMLAdapterTests):
         del mh.smtp_queue
         del mh.smtp_queue_directory
         self._BODY = _MAILHOST_BODY
+        zope.component.provideAdapter(self._getTargetClass())
 
 
 @unittest.skipUnless(HAVE_GS, 'Products.GenericSetup not available.')
@@ -103,6 +111,8 @@ class MailHostXMLAdapterTestsWithQueue(BodyAdapterTestCase, unittest.TestCase):
 
         self._obj = MailHost('foo_mailhost')
         self._BODY = _MAILHOST_BODY_v2
+        zope.component.provideAdapter(self._getTargetClass())
+
 
 
 @unittest.skipUnless(HAVE_GS, 'Products.GenericSetup not available.')
@@ -120,6 +130,7 @@ class MailHostXMLAdapterTestsWithNoneValue(MailHostXMLAdapterTests):
         self._obj = MailHost('foo_mailhost')
         self._obj.smtp_uid = None
         self._BODY = _MAILHOST_BODY
+        zope.component.provideAdapter(self._getTargetClass())
 
 
 def test_suite():