+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog
 4.2 (unreleased)
 ----------------
 
+- Fix tests to run with `Products.GenericSetup >= 1.10`.
 
 4.1 (2018-05-20)
 ----------------